requirements.txt
djangorestframework==3.9.2
Django==2.1.7
requests